+++ examples/demo_project/config/config.exs
@@ -1,3 +1,1 @@
 import Config
-
-config :mneme, dry_run: true